Now it is possible to have multiple stable objects such as Reithalle, Roundpen, Reitplatz or whatever. For this, we needed to extend the database structure with a new database table called "stable_object" containing an unique id, also the stable id which the object belongs to and a name for the stable object. Those information are now available within events table, which a new column called stable_object_id which refers to the stable object. those data are now also logged,too.
Registration process:
We did not change much but when a new stable is registered we need an object per default such as "Reithalle".
So now the stable is inserted during the registration process and immediatelly creates a new entry within stable_object table.
PEr default this is the only calender.
Admins can not enter a new stable object on their own. This must be done by backend administration for now.

So ... the whole process is working now.

Also, changed the "my entries" overview. Not the time formats are identical in the "past" and "future" dates as well as the "Nr." tag.
The last entry in the overview column is no longer the event title but the object...so users can see where they booked.